cleaned up detect
